Extract vitest.config into special folder under extensions
